Renamed kandinsky module, fixed img dimensions description
