+++ bot/chat_modes.json
@@ -25,7 +25,7 @@
   },
     "image": {
     "name": "\uD83D\uDDBC️ Image",
-    "welcome_message": "\uD83D\uDDBC️ <b>Kandinsky 2.2</b>. It will generate a 512x512 image based on a prompt.",
+    "welcome_message": "\uD83D\uDDBC️ <b>Kandinsky 2.2</b>. It will generate a 1024x1024 image based on a prompt.",
     "prompt_start": "",
     "parse_mode": "html"
   }
