+++ bot/config.py
@@ -16,7 +16,6 @@ telegram_token = config_yaml["telegram_token"]
 openai_login = config_yaml["openai_login"]
 openai_password = config_yaml["openai_password"]
 openai_access_token = config_yaml["openai_access_token"]
-replicate_api_token = config_yaml["replicate_api_token"]
 allowed_telegram_usernames = config_yaml["allowed_telegram_usernames"]
 new_dialog_timeout = config_yaml["new_dialog_timeout"]
 enable_message_streaming = config_yaml.get("enable_message_streaming", True)
